requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
idna==3.7
mysql-connector-python==9.0.0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
